Refactor Slurm REST client
