+++ src/cloudai/systems/slurm/__init__.py
@@ -20,8 +20,9 @@ from .slurm_installer import SlurmInstaller
 from .slurm_job import SlurmJob
 from .slurm_metadata import SlurmJobMetadata, SlurmStepMetadata, SlurmSystemMetadata
 from .slurm_node import SlurmNode, SlurmNodeState
+from .slurm_rest_client import SlurmAPIConfig
 from .slurm_runner import SlurmRunner
-from .slurm_system import SlurmAPIConfig, SlurmGroup, SlurmPartition, SlurmSystem, parse_node_list
+from .slurm_system import SlurmGroup, SlurmPartition, SlurmSystem, parse_node_list
 
 __all__ = [
     "SingleSbatchRunner",
