+++ tests/systems/slurm/test_system.py
@@ -35,6 +35,7 @@ from cloudai.systems.slurm import (
     parse_node_list,
 )
 from cloudai.systems.slurm.slurm_metadata import SlurmStepMetadata
+from cloudai.systems.slurm.slurm_rest_client import SlurmRestClient
 from cloudai.workloads.nccl_test import NCCLCmdArgs, NCCLTestDefinition
 
 
@@ -57,8 +58,8 @@ def test_slurm_api_request_expands_headers(rest_slurm_system: SlurmSystem, monke
     response = Mock()
     response.json.return_value = {"pings": [], "errors": [], "warnings": []}
 
-    with patch("cloudai.systems.slurm.slurm_system.requests.request", return_value=response) as request:
-        rest_slurm_system._rest_request("GET", "slurm", "ping/")
+    with patch("cloudai.systems.slurm.slurm_rest_client.requests.request", return_value=response) as request:
+        rest_slurm_system._rest_client.request("GET", "slurm", "ping/")
 
     request.assert_called_once_with(
         "GET",
@@ -93,7 +94,7 @@ srun --container-image=image.sqsh benchmark
     script_path = tmp_path / "job.sbatch"
     script_path.write_text(script)
 
-    with patch.object(rest_slurm_system, "_rest_request", return_value={"job_id": 123}) as request:
+    with patch.object(SlurmRestClient, "request", return_value={"job_id": 123}) as request:
         job_id = rest_slurm_system.submit_job(f"sbatch {script_path}", "rest-test")
 
     assert job_id == 123
@@ -123,7 +124,7 @@ def test_slurm_api_rejects_unsupported_sbatch_directive(rest_slurm_system: Slurm
     script_path.write_text("#!/bin/bash\n#SBATCH --qos=high\nsrun true\n")
 
     with (
-        patch.object(rest_slurm_system, "_rest_request") as request,
+        patch.object(SlurmRestClient, "request") as request,
         pytest.raises(JobIdRetrievalError, match="Failed to submit job through Slurm REST API"),
     ):
         rest_slurm_system.submit_sbatch(script_path, "rest-test")
@@ -159,7 +160,7 @@ def test_slurm_api_job_lifecycle(rest_slurm_system: SlurmSystem):
     }
     job = SlurmJob(test_run=Mock(), id=42)
 
-    with patch.object(rest_slurm_system, "_rest_request", return_value=response):
+    with patch.object(SlurmRestClient, "request", return_value=response):
         assert rest_slurm_system.is_job_running(job) is False
         assert rest_slurm_system.is_job_completed(job) is True
         assert rest_slurm_system.complete_job(job) == ["node01", "node02"]
@@ -201,7 +202,7 @@ def test_slurm_api_nodes_cancel_and_validation(rest_slurm_system: SlurmSystem):
 
     rest_slurm_system.supports_gpu_directives_cache = None
     with (
-        patch.object(rest_slurm_system, "_rest_request", side_effect=request) as rest_request,
+        patch.object(SlurmRestClient, "request", side_effect=request) as rest_request,
         patch("cloudai.systems.slurm.slurm_system.shutil.which", return_value="/usr/bin/git"),
     ):
         assert rest_slurm_system.supports_gpu_directives is True
